Update README.md
I have and always will be looking for new unconventional ways to gain inspiration and or a muse of some type, generally in the form of a beautiful woman, in order to grow closer to reaching my greatest potential which I personally believe is very much more than what my life currently portrays. Doing so has and always will be a challenge with all the demands that life has to offer for us to overcome and conquer if we want to have a good one, and the dedication to better ones situation should reflect in one's work- being a benefit to both their quality of life, and therefore their output.